Move comlink worker mock to den (#7357)
This same pattern can be applied when testing any class that uses a
worker interface via comlink. By putting this in den we can re-use this
pattern instead of copy/pasting it.
