+++ packages/den/package.json
@@ -14,6 +14,7 @@
   "homepage": "https://foxglove.dev/",
   "dependencies": {
     "async-mutex": "0.4.0",
+    "comlink": "github:foxglove/comlink#216a78c5fb08550fdef4a725f468999707fec5bd",
     "xacro-parser": "0.3.9"
   },
   "devDependencies": {
